feat(#682): improved reusage for create orchestrator
