+++ packages/core/src/governance/client/hooks/useOrchestratorTasks.ts
@@ -1,0 +1,135 @@
+'use client';
+
+import { useCallback, useReducer, useState } from 'react';
+import { produce } from 'immer';
+import { useAgreementFileUploads } from './useAgreementFileUploads';
+import { useAgreementMutationsWeb2Rsc } from './useAgreementMutations.web2.rsc';
+
+export type TaskName =
+  | 'CREATE_WEB2_AGREEMENT'
+  | 'CREATE_WEB3_AGREEMENT'
+  | 'UPLOAD_FILES'
+  | 'LINK_WEB2_AND_WEB3_AGREEMENT';
+
+export type TaskState = {
+  [K in TaskName]: {
+    status: TaskStatus;
+    message?: string;
+  };
+};
+
+export enum TaskStatus {
+  IDLE = 'idle',
+  IS_PENDING = 'isPending',
+  IS_DONE = 'isDone',
+  ERROR = 'error',
+}
+
+export const taskActionDescriptions: Record<TaskName, string> = {
+  CREATE_WEB2_AGREEMENT: 'Creating Web2 agreement...',
+  CREATE_WEB3_AGREEMENT: 'Creating Web3 agreement...',
+  UPLOAD_FILES: 'Uploading Agreement Files...',
+  LINK_WEB2_AND_WEB3_AGREEMENT: 'Linking Web2 and Web3 agreements',
+};
+
+export const initialTaskState: TaskState = {
+  CREATE_WEB2_AGREEMENT: { status: TaskStatus.IDLE },
+  CREATE_WEB3_AGREEMENT: { status: TaskStatus.IDLE },
+  UPLOAD_FILES: { status: TaskStatus.IDLE },
+  LINK_WEB2_AND_WEB3_AGREEMENT: { status: TaskStatus.IDLE },
+};
+
+export type ProgressAction =
+  | { type: 'START_TASK'; taskName: TaskName; message?: string }
+  | { type: 'COMPLETE_TASK'; taskName: TaskName; message?: string }
+  | { type: 'SET_ERROR'; taskName: TaskName; message: string }
+  | { type: 'RESET' };
+
+export const progressStateReducer = (
+  state: TaskState,
+  action: ProgressAction,
+): TaskState =>
+  produce(state, (draft) => {
+    switch (action.type) {
+      case 'START_TASK':
+        draft[action.taskName].status = TaskStatus.IS_PENDING;
+        draft[action.taskName].message = action.message;
+        break;
+      case 'COMPLETE_TASK':
+        draft[action.taskName].status = TaskStatus.IS_DONE;
+        draft[action.taskName].message = action.message;
+        break;
+      case 'SET_ERROR':
+        draft[action.taskName].status = TaskStatus.ERROR;
+        draft[action.taskName].message = action.message;
+        break;
+      case 'RESET':
+        return initialTaskState;
+    }
+  });
+
+export const computeProgress = (tasks: TaskState): number => {
+  const all = Object.values(tasks);
+  const done = all.filter((t) => t.status === TaskStatus.IS_DONE).length;
+  const pending = all.filter((t) => t.status === TaskStatus.IS_PENDING).length;
+  return Math.round(((done + pending * 0.5) / all.length) * 100);
+};
+
+export function useOrchestratorTasks({
+  authToken,
+}: {
+  authToken?: string | null;
+}) {
+  const agreementFiles = useAgreementFileUploads(authToken);
+  const web2 = useAgreementMutationsWeb2Rsc(authToken);
+
+  const [taskState, dispatch] = useReducer(
+    progressStateReducer,
+    initialTaskState,
+  );
+  const [currentAction, setCurrentAction] = useState<string>();
+
+  const progress = computeProgress(taskState);
+
+  const startTask = useCallback((taskName: TaskName) => {
+    const message = taskActionDescriptions[taskName];
+    setCurrentAction(message);
+    dispatch({ type: 'START_TASK', taskName, message });
+  }, []);
+
+  const completeTask = useCallback(
+    (taskName: TaskName) => {
+      if (currentAction === taskActionDescriptions[taskName]) {
+        setCurrentAction(undefined);
+      }
+      dispatch({ type: 'COMPLETE_TASK', taskName });
+    },
+    [currentAction],
+  );
+
+  const errorTask = useCallback(
+    (taskName: TaskName, error: string) => {
+      if (currentAction === taskActionDescriptions[taskName]) {
+        setCurrentAction(undefined);
+      }
+      dispatch({ type: 'SET_ERROR', taskName, message: error });
+    },
+    [currentAction],
+  );
+
+  const resetTasks = useCallback(() => {
+    dispatch({ type: 'RESET' });
+  }, []);
+
+  return {
+    taskState,
+    currentAction,
+    agreementFiles,
+    web2,
+    progress,
+    startTask,
+    completeTask,
+    errorTask,
+    resetTasks,
+  };
+}
